[DDW-69] Improve exporter code
